Refactored page ts files
